parse authentication credentials from url for http and ftp schemes
